+++ Source/CNTKv2LibraryDll/Function.cpp
@@ -1061,7 +1061,7 @@ namespace CNTK
         }
     }
 
-    FunctionPtr EditDistanceError(const Variable& prediction, const Variable& labels, float subPen, float delPen, float insPen, bool squashInputs, const vector<int>& samplesToIgnore, const std::wstring& name)
+    FunctionPtr EditDistanceError(const Variable& prediction, const Variable& labels, float subPen, float delPen, float insPen, bool squashInputs, const vector<size_t>& samplesToIgnore, const std::wstring& name)
     {
         auto additionalProperties = Dictionary();
         additionalProperties[PrimitiveFunction::AttributeNameSubstitutionPenalty] = subPen;
